<update>: riscv64-linux-gnu-gcc can use rvld as linker. And can print all args from riscv64-linux-gnu-gcc.

+++ src/InputFile.cpp
@@ -41,7 +41,7 @@ vector<uint8_t> InputFile::GetBytesFromIdx(int idx) {
 }
 
 Shdr* InputFile::FindSection(uint32_t type) {
-    for(int i=0; i<ElfSections.size(); i++) {
+    for(size_t i=0; i<ElfSections.size(); i++) {
         Shdr* s = &ElfSections[i];
         if(s->Type == type)
             return s;
